Using short instead of int

+++ src/java/is/idega/idegaweb/egov/bpm/cases/presentation/beans/CasesSearchResultsHolderImpl.java
@@ -253,14 +253,14 @@ public class CasesSearchResultsHolderImpl implements CasesSearchResultsHolder {
 			List<String> standardFieldsInfo) {
 		IWResourceBundle iwrb = getResourceBundle(CasesConstants.IW_BUNDLE_IDENTIFIER);
 		
-		int cellIndexInRow = 0;
+		short cellIndexInRow = 0;
 		
 		sheet.setColumnWidth(cellIndexInRow++, DEFAULT_CELL_WIDTH);
 		sheet.setColumnWidth(cellIndexInRow++, DEFAULT_CELL_WIDTH);
 		sheet.setColumnWidth(cellIndexInRow++, DEFAULT_CELL_WIDTH);
 		
-		int cellRow = 0;
-		int cellIndex = 0;
+		short cellRow = 0;
+		short cellIndex = 0;
 		
 		//	Default header labels
 		HSSFRow row = sheet.createRow(cellRow++);
@@ -334,7 +334,7 @@ public class CasesSearchResultsHolderImpl implements CasesSearchResultsHolder {
 	}
 	
 	private void addVariables(List<AdvancedProperty> variablesByProcessDefinition, CasePresentation theCase, HSSFRow row, HSSFSheet sheet, HSSFCellStyle bigStyle,
-			Locale locale, boolean isAdmin, int cellIndex, List<Integer> fileCellsIndexes, String localizedFileLabel) {
+			Locale locale, boolean isAdmin, short cellIndex, List<Integer> fileCellsIndexes, String localizedFileLabel) {
 		if (ListUtil.isEmpty(variablesByProcessDefinition)) {
 			return;
 		}
@@ -411,12 +411,12 @@ public class CasesSearchResultsHolderImpl implements CasesSearchResultsHolder {
 			
 			List<AdvancedProperty> variablesByProcessDefinition = createHeaders(sheet, bigStyle, locale, processName, isAdmin, standardFieldsLabels);
 			List<Integer> fileCellsIndexes = null;
-			int rowNumber = 1;
+			short rowNumber = 1;
 			
 			for (CasePresentation theCase: cases) {
 				fileCellsIndexes = new ArrayList<Integer>();
 				HSSFRow row = sheet.createRow(rowNumber++);
-				int cellIndex = 0;
+				short cellIndex = 0;
 
 				//	Default header values
 				row.createCell(cellIndex++).setCellValue(theCase.getCaseIdentifier());
